refactor(core): split ModuleMap into focused submodules
